refactor unittest
